requirements.txt
# BlueForecast Data Pipeline Dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyarrow>=12.0.0
google-cloud-storage>=2.10.0
requests>=2.28.0
apache-airflow>=2.7.0
dvc[gs]>=3.0.0
pytest>=7.0.0